+++ config/local.ts
@@ -1,6 +1,8 @@
 import {Config, ForeignConfig, GlobalConfig} from './types';
 import {readValidators} from './utils';
 
+const replicator = '0x620583C75BB474E06485893B795b0883b5816D10';
+
 const config: Config = {
   contractRegistry: {
     address: '0x7c2C195CD6D34B8F845992d380aADB2730bB9C6F'
@@ -13,7 +15,7 @@ const config: Config = {
   chain: {
     padding: 20,
     requiredSignatures: 1,
-    replicator: '0x620583C75BB474E06485893B795b0883b5816D10'
+    replicator
   },
   stakingBank: {
     minAmountForStake: 1n * BigInt(1e18),
@@ -28,7 +30,7 @@ const foreignConfig: ForeignConfig = {
   chain: {
     padding: 20,
     requiredSignatures: 1,
-    replicator: '0x620583C75BB474E06485893B795b0883b5816D10'
+    replicator
   },
 };
 
@@ -36,6 +38,7 @@ const local: GlobalConfig = {
   bsc: config,
   ethereum: foreignConfig,
   polygon: foreignConfig,
+  avalanche: foreignConfig,
 };
 
 export default local;