+++ config/sandbox.ts
@@ -1,7 +1,22 @@
 import {GlobalConfig} from './types';
 import {readValidators} from './utils';
 
+const replicator = '0x620583c75bb474e06485893b795b0883b5816d10';
+
 const staging: GlobalConfig = {
+  avalanche: {
+    distributor: {
+      address: '0xbC758FCB97e06Ec635DFf698f55e41aCC35e1d2d'
+    },
+    contractRegistry: {
+      address: '0x26fD86791fCE0946E8D8c685446Dd257634a2b28'
+    },
+    chain: {
+      padding: 60,
+      requiredSignatures: 1,
+      replicator
+    },
+  },
   polygon: {
     distributor: {
       address: '0x3fBdba5E516611e2E9C1aA6cdFB8376c7cA7d7CE'
@@ -12,7 +27,7 @@ const staging: GlobalConfig = {
     chain: {
       padding: 60,
       requiredSignatures: 1,
-      replicator: '0x620583c75bb474e06485893b795b0883b5816d10'
+      replicator
     },
   },
   ethereum: {
@@ -25,7 +40,7 @@ const staging: GlobalConfig = {
     chain: {
       padding: 300,
       requiredSignatures: 1,
-      replicator: '0x620583c75bb474e06485893b795b0883b5816d10'
+      replicator
     },
   },
   bsc: {
@@ -43,7 +58,7 @@ const staging: GlobalConfig = {
     chain: {
       padding: 20,
       requiredSignatures: 1,
-      replicator: '0x0760b24EC17dDf59893D7c16D75353dE24c00F14'
+      replicator
     },
     stakingBank: {
       minAmountForStake: 100n * BigInt(1e18),