+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `Distributor` contract
+
+### Removed
+- local minter, as hardhat can finally mint
 
 ## [4.2.0] - 2021-10-18
 ### Added